Rename test files and update import paths for consistency in food and user components
